requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Werkzeug==2.3.7
azure-storage-blob==12.19.0
python-dotenv==1.0.0
WTForms==3.0.1
Flask-WTF==1.1.1
email-validator==2.0.0